requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.4
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
docker==7.1.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
isort==5.13.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-anthropic==0.1.21
langchain-core==0.2.24
langchain-text-splitters==0.2.2
langsmith==0.1.93
mccabe==0.7.0
multidict==6.0.5
mypy==1.11.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
tenacity==8.5.0
tokenizers==0.19.1
tomlkit==0.13.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
